+++ system/src/Grav/Common/Page/Interfaces/PageRoutableInterface.php
@@ -63,7 +63,7 @@ interface PageRoutableInterface
      * the parents route and the current Page's slug.
      *
      * @param  string|null $var Set new default route.
-     * @return string  The route for the Page.
+     * @return string|null  The route for the Page.
      */
     public function route($var = null);