+++ system/src/Grav/Common/Theme.php
@@ -9,9 +9,9 @@
 
 namespace Grav\Common;
 
-use Grav\Common\Page\Interfaces\PageInterface;
 use Grav\Common\Config\Config;
 use RocketTheme\Toolbox\File\YamlFile;
+use RocketTheme\Toolbox\ResourceLocator\UniformResourceLocator;
 
 /**
  * Class Theme
@@ -44,51 +44,28 @@ class Theme extends Plugin
     /**
      * Persists to disk the theme parameters currently stored in the Grav Config object
      *
-     * @param string $theme_name The name of the theme whose config it should store.
+     * @param string $name The name of the theme whose config it should store.
      * @return bool
      */
-    public static function saveConfig($theme_name)
+    public static function saveConfig($name)
     {
-        if (!$theme_name) {
+        if (!$name) {
             return false;
         }
 
         $grav = Grav::instance();
+
+        /** @var UniformResourceLocator $locator */
         $locator = $grav['locator'];
-        $filename = 'config://themes/' . $theme_name . '.yaml';
-        $file = YamlFile::instance($locator->findResource($filename, true, true));
-        $content = $grav['config']->get('themes.' . $theme_name);
+
+        $filename = 'config://themes/' . $name . '.yaml';
+        $file = YamlFile::instance((string)$locator->findResource($filename, true, true));
+        $content = $grav['config']->get('themes.' . $name);
         $file->save($content);
         $file->free();
+        unset($file);
 
         return true;
-    }
-
-    /**
-     * Override the mergeConfig method to work for themes
-     *
-     * @param PageInterface $page
-     * @param string $deep
-     * @param array $params
-     * @param string $type
-     * @return Data\Data
-     */
-    protected function mergeConfig(PageInterface $page, $deep = 'merge', $params = [], $type = 'themes')
-    {
-        return parent::mergeConfig($page, $deep, $params, $type);
-    }
-
-    /**
-     * Simpler getter for the theme blueprint
-     *
-     * @return mixed
-     */
-    public function getBlueprint()
-    {
-        if (!$this->blueprint) {
-            $this->loadBlueprint();
-        }
-        return $this->blueprint;
     }
 
     /**
@@ -100,8 +77,11 @@ class Theme extends Plugin
     {
         if (!$this->blueprint) {
             $grav = Grav::instance();
+            /** @var Themes $themes */
             $themes = $grav['themes'];
-            $this->blueprint = $themes->get($this->name)->blueprints();
+            $data = $themes->get($this->name);
+            \assert($data !== null);
+            $this->blueprint = $data->blueprints();
         }
     }
 }