+++ system/src/Grav/Framework/Flex/Pages/Traits/PageRoutableTrait.php
@@ -18,6 +18,7 @@ use Grav\Common\Uri;
 use Grav\Framework\Filesystem\Filesystem;
 use RocketTheme\Toolbox\ResourceLocator\UniformResourceLocator;
 use RuntimeException;
+use function dirname;
 use function is_string;
 
 /**
@@ -32,6 +33,8 @@ trait PageRoutableTrait
     private $_route;
     /** @var string|null */
     private $_path;
+    /** @var PageInterface|null */
+    private $_parentCache;
 
     /**
      * Returns the page extension, got from the page `url_extension` config and falls back to the
@@ -317,7 +320,7 @@ trait PageRoutableTrait
 
         /** @var UniformResourceLocator $locator */
         $locator = Grav::instance()['locator'];
-        $path = $locator->findResource($folder, false);
+        $path = $locator->isStream($folder) ? $locator->findResource($folder, false) : $folder;
 
         return is_string($path) ? $path : null;
     }
@@ -350,7 +353,7 @@ trait PageRoutableTrait
         if ($folder) {
             /** @var UniformResourceLocator $locator */
             $locator = Grav::instance()['locator'];
-            $folder = $locator($folder);
+            $folder = $locator->isStream($folder) ? $locator->getResource($folder) : GRAV_ROOT . "/{$folder}";
         }
 
         return $this->_path = is_string($folder) ? $folder : null;
@@ -413,13 +416,12 @@ trait PageRoutableTrait
             throw new RuntimeException(__METHOD__ . '(PageInterface): Not Implemented');
         }
 
-        if ($this->root()) {
-            return null;
+        if ($this->_parentCache || $this->root()) {
+            return $this->_parentCache;
         }
 
-        $filesystem = Filesystem::getInstance(false);
         $directory = $this->getFlexDirectory();
-        $parentKey = ltrim($filesystem->dirname("/{$this->getKey()}"), '/');
+        $parentKey = ltrim(dirname("/{$this->getKey()}"), '/');
         if ($parentKey) {
             $parent = $directory->getObject($parentKey);
             $language = $this->getLanguage();
@@ -427,12 +429,14 @@ trait PageRoutableTrait
                 $parent = $parent->getTranslation($language) ?? $parent;
             }
 
-            return $parent;
+            $this->_parentCache = $parent;
+        } else {
+            $index = $directory->getIndex();
+
+            $this->_parentCache = method_exists($index, 'getRoot') ? $index->getRoot() : null;
         }
 
-        $index = $directory->getIndex();
-
-        return method_exists($index, 'getRoot') ? $index->getRoot() : null;
+        return $this->_parentCache;
     }
 
     /**