+++ system/src/Grav/Framework/Flex/Storage/FolderStorage.php
@@ -377,12 +377,14 @@ class FolderStorage extends AbstractFilesystemStorage
         $file = $this->getFile($path);
         try {
             $data = (array)$file->content();
-            $file->free();
             if (isset($data[0])) {
                 throw new RuntimeException('Broken object file');
             }
         } catch (RuntimeException $e) {
             $data = ['__ERROR' => $e->getMessage()];
+        } finally {
+            $file->free();
+            unset($file);
         }
 
         $data['__META'] = $this->getObjectMeta($key);
@@ -426,13 +428,17 @@ class FolderStorage extends AbstractFilesystemStorage
 
             $file->save($row);
 
-            /** @var UniformResourceLocator $locator */
-            $locator = Grav::instance()['locator'];
-            if ($locator->isStream($path)) {
-                $locator->clearCache();
-            }
         } catch (RuntimeException $e) {
             throw new RuntimeException(sprintf('Flex saveFile(%s): %s', $path ?? $key, $e->getMessage()));
+        } finally {
+            /** @var UniformResourceLocator $locator */
+            $locator = Grav::instance()['locator'];
+            $locator->clearCache();
+
+            if (isset($file)) {
+                $file->free();
+                unset($file);
+            }
         }
 
         $row['__META'] = $this->getObjectMeta($key, true);
@@ -452,14 +458,14 @@ class FolderStorage extends AbstractFilesystemStorage
             if ($file->exists()) {
                 $file->delete();
             }
-
-            /** @var UniformResourceLocator $locator */
-            $locator = Grav::instance()['locator'];
-            if ($locator->isStream($filename)) {
-                $locator->clearCache($filename);
-            }
         } catch (RuntimeException $e) {
             throw new RuntimeException(sprintf('Flex deleteFile(%s): %s', $filename, $e->getMessage()));
+        } finally {
+            /** @var UniformResourceLocator $locator */
+            $locator = Grav::instance()['locator'];
+            $locator->clearCache();
+
+            $file->free();
         }
 
         return $data;
@@ -474,14 +480,12 @@ class FolderStorage extends AbstractFilesystemStorage
     {
         try {
             Folder::copy($this->resolvePath($src), $this->resolvePath($dst));
-
-            /** @var UniformResourceLocator $locator */
-            $locator = Grav::instance()['locator'];
-            if ($locator->isStream($src) || $locator->isStream($dst)) {
-                $locator->clearCache();
-            }
         } catch (RuntimeException $e) {
             throw new RuntimeException(sprintf('Flex copyFolder(%s, %s): %s', $src, $dst, $e->getMessage()));
+        } finally {
+            /** @var UniformResourceLocator $locator */
+            $locator = Grav::instance()['locator'];
+            $locator->clearCache();
         }
 
         return true;
@@ -496,14 +500,12 @@ class FolderStorage extends AbstractFilesystemStorage
     {
         try {
             Folder::move($this->resolvePath($src), $this->resolvePath($dst));
-
-            /** @var UniformResourceLocator $locator */
-            $locator = Grav::instance()['locator'];
-            if ($locator->isStream($src) || $locator->isStream($dst)) {
-                $locator->clearCache();
-            }
         } catch (RuntimeException $e) {
             throw new RuntimeException(sprintf('Flex moveFolder(%s, %s): %s', $src, $dst, $e->getMessage()));
+        } finally {
+            /** @var UniformResourceLocator $locator */
+            $locator = Grav::instance()['locator'];
+            $locator->clearCache();
         }
 
         return true;
@@ -517,17 +519,13 @@ class FolderStorage extends AbstractFilesystemStorage
     protected function deleteFolder(string $path, bool $include_target = false): bool
     {
         try {
-            $success = Folder::delete($this->resolvePath($path), $include_target);
-
-            /** @var UniformResourceLocator $locator */
-            $locator = Grav::instance()['locator'];
-            if ($locator->isStream($path)) {
-                $locator->clearCache();
-            }
-
-            return $success;
+            return Folder::delete($this->resolvePath($path), $include_target);
         } catch (RuntimeException $e) {
             throw new RuntimeException(sprintf('Flex deleteFolder(%s): %s', $path, $e->getMessage()));
+        } finally {
+            /** @var UniformResourceLocator $locator */
+            $locator = Grav::instance()['locator'];
+            $locator->clearCache();
         }
     }
 
@@ -669,7 +667,14 @@ class FolderStorage extends AbstractFilesystemStorage
         /** @var string $pattern */
         $pattern = !empty($options['pattern']) ? $options['pattern'] : $this->dataPattern;
 
-        $this->dataFolder = $options['folder'];
+        /** @var UniformResourceLocator $locator */
+        $locator = Grav::instance()['locator'];
+        $folder = $options['folder'];
+        if ($locator->isStream($folder)) {
+            $folder = $locator->getResource($folder, false);
+        }
+
+        $this->dataFolder = $folder;
         $this->dataFile = $options['file'] ?? 'item';
         $this->dataExt = $extension;
         if (mb_strpos($pattern, '{FILE}') === false && mb_strpos($pattern, '{EXT}') === false) {